pipe: Introduce Buffer type for message buffering
